Fixed warnings about type-conversion.

+++ samples/deeplearning/rnndriver/rnndriver.c
@@ -154,7 +154,7 @@ LIBXSMM_INLINE void matrix_relu_inverse(int size, float *src, float *dst)
 # pragma omp parallel for private(i)
 #endif
   for (i = 0; i < size; i++) {
-    dst[i] = (src[i] > 0) ? 1 : 0;
+    dst[i] = (float)(src[i] > 0 ? 1 : 0);
   }
 }
